requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch
torchvision
Pillow
requests
python-dotenv
pytest
httpx
# Removed google-generativeai as it is not used in the current code.
